+++ src/Toolkit/kits/bootstrap/tooltip/assets/controllers/tooltip_controller.js
@@ -1,0 +1,12 @@
+import { Controller } from '@hotwired/stimulus';
+import { Tooltip } from 'bootstrap';
+
+export default class extends Controller {
+    connect() {
+        this._tooltip = new Tooltip(this.element);
+    }
+
+    disconnect() {
+        this._tooltip?.dispose();
+    }
+}