+++ src/Toolkit/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.35
 
+- [Flowbite] Add Flowbite v4 kit
 - [Shadcn] Add `toggle` recipe
 - [Shadcn] Use `html_attr_type` filter from `twig/html-extra:^3.24` for composable trigger attributes
 - [Shadcn] Rename `trigger_attrs` to `alert_dialog_trigger_attrs` in `AlertDialog:Trigger`