+++ src/Toolkit/CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ## 2.35
 
+- [Flowbite] Add Flowbite v4 kit
+- [Shadcn] Add `toggle` recipe
 - [Shadcn] Use `html_attr_type` filter from `twig/html-extra:^3.24` for composable trigger attributes
 - [Shadcn] Rename `trigger_attrs` to `alert_dialog_trigger_attrs` in `AlertDialog:Trigger`
 - [Shadcn] Rename `trigger_attrs` to `dialog_trigger_attrs` in `Dialog:Trigger`