+++ src/Toolkit/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Unreleased
+
+- [Bootstrap] Add Bootstrap kit with
+
 ## 3.0.0
 
 - Minimum required Symfony version is now 7.4