+++ src/Toolkit/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Minimum required Symfony version is now 7.4
 - Minimum required PHP version is now 8.4
+- [Shadcn] Add `input-otp` recipe
 
 ## 2.35