+++ src/Toolkit/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Minimum required Symfony version is now 7.4
 - Minimum required PHP version is now 8.4
 - [Shadcn] Add `menubar` recipe
+- [Shadcn] Add `popover` recipe
 
 ## 2.35